Add "references" parameter to TestHelpers.GetDocument()
