--- Source/RoslynNUnitLight/MarkupCode.cs
@@ -1,38 +1,0 @@
-﻿using System.Text;
-using Microsoft.CodeAnalysis.Text;
-
-namespace RoslynNUnitLight
-{
-    public static class MarkupCode
-    {
-        public static bool TryGetCodeAndSpan(string markupCode, out string code, out TextSpan span)
-        {
-            code = null;
-            span = default(TextSpan);
-
-            var builder = new StringBuilder();
-
-            var start = markupCode.IndexOf("[|");
-            if (start < 0)
-            {
-                return false;
-            }
-
-            builder.Append(markupCode.Substring(0, start));
-
-            var end = markupCode.IndexOf("|]");
-            if (end < 0)
-            {
-                return false;
-            }
-
-            builder.Append(markupCode.Substring(start + 2, end - start - 2));
-            builder.Append(markupCode.Substring(end + 2));
-
-            code = builder.ToString();
-            span = TextSpan.FromBounds(start, end - 2);
-
-            return true;
-        }
-    }
-}
