+++ Source/RoslynNUnitLight/TestHelpers.cs
@@ -1,0 +1,67 @@
+﻿using System.Collections.Immutable;
+using System.Linq;
+using System.Reflection;
+using System.Text;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.Text;
+
+namespace RoslynNUnitLight
+{
+    public static class Code
+    {
+        public static bool TryGetCodeAndSpanFromMarkup(string markupCode, out string code, out TextSpan span)
+        {
+            code = null;
+            span = default(TextSpan);
+
+            var builder = new StringBuilder();
+
+            var start = markupCode.IndexOf("[|");
+            if (start < 0)
+            {
+                return false;
+            }
+
+            builder.Append(markupCode.Substring(0, start));
+
+            var end = markupCode.IndexOf("|]");
+            if (end < 0)
+            {
+                return false;
+            }
+
+            builder.Append(markupCode.Substring(start + 2, end - start - 2));
+            builder.Append(markupCode.Substring(end + 2));
+
+            code = builder.ToString();
+            span = TextSpan.FromBounds(start, end - 2);
+
+            return true;
+        }
+
+        public static bool TryGetDocumentAndSpanFromMarkup(string markupCode, string languageName, out Document document, out TextSpan span)
+        {
+            string code;
+            if (!TryGetCodeAndSpanFromMarkup(markupCode, out code, out span))
+            {
+                document = null;
+                return false;
+            }
+
+            document = GetDocument(code, languageName);
+            return true;
+        }
+
+        internal static Document GetDocument(string code, string languageName, ImmutableList<MetadataReference> references = null)
+        {
+            references = references ?? ImmutableList.Create(
+                MetadataReference.CreateFromAssembly(typeof(object).GetTypeInfo().Assembly),
+                MetadataReference.CreateFromAssembly(typeof(Enumerable).GetTypeInfo().Assembly));
+
+            return new AdhocWorkspace()
+                .AddProject("TestProject", languageName)
+                .AddMetadataReferences(references)
+                .AddDocument("TestDocument", code);
+        }
+    }
+}
